Image generate on dialog (#177)
* image generation on save, css improvements, cleanup
